requirements.txt
blis==0.4.1
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cymem==2.0.3
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
Flask==1.1.2
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
murmurhash==1.0.2
numpy==1.19.1
opencv-python==4.3.0.36
Pillow==7.2.0
plac==1.1.3
preshed==3.0.2
pycountry==20.7.3
pytesseract==0.3.4
requests==2.24.0
spacy==2.3.2
srsly==1.0.2
thinc==7.4.1
tqdm==4.48.2
urllib3==1.25.10
wasabi==0.7.1
Werkzeug==1.0.1
zipp==3.1.0
